Refactoring of the connection params
This minor refactoring is a pre-requisition of the following extraction
of the logic of parsing a gRPC metadata.
