+++ grpc/params.go
@@ -31,7 +31,7 @@ func newCallParams(vu modules.VU, input goja.Value) (*callParams, error) {
 		TagsAndMeta: vu.State().Tags.GetCurrentValues(),
 	}
 
-	if input == nil || goja.IsUndefined(input) || goja.IsNull(input) {
+	if common.IsNullish(input) {
 		return result, nil
 	}
 
@@ -46,6 +46,7 @@ func newCallParams(vu modules.VU, input goja.Value) (*callParams, error) {
 			if !ok {
 				return result, errors.New("metadata must be an object with key-value pairs")
 			}
+
 			for hk, kv := range rawHeaders {
 				var val string
 
@@ -98,4 +99,122 @@ func (p *callParams) SetSystemTags(state *lib.State, addr string, methodName str
 	if _, ok := p.TagsAndMeta.Tags.Get("name"); !ok {
 		p.TagsAndMeta.SetSystemTagOrMetaIfEnabled(state.Options.SystemTags, metrics.TagName, methodName)
 	}
+}
+
+// connectParams is the parameters that can be passed to a gRPC connect call.
+type connectParams struct {
+	IsPlaintext           bool
+	UseReflectionProtocol bool
+	Timeout               time.Duration
+	MaxReceiveSize        int64
+	MaxSendSize           int64
+	TLS                   map[string]interface{}
+}
+
+func newConnectParams(vu modules.VU, input goja.Value) (*connectParams, error) {
+	result := &connectParams{
+		IsPlaintext:           false,
+		UseReflectionProtocol: false,
+		Timeout:               time.Minute,
+		MaxReceiveSize:        0,
+		MaxSendSize:           0,
+	}
+
+	if common.IsNullish(input) {
+		return result, nil
+	}
+
+	rt := vu.Runtime()
+	params := input.ToObject(rt)
+
+	for _, k := range params.Keys() {
+		v := params.Get(k).Export()
+
+		switch k {
+		case "plaintext":
+			var ok bool
+			result.IsPlaintext, ok = v.(bool)
+			if !ok {
+				return result, fmt.Errorf("invalid plaintext value: '%#v', it needs to be boolean", v)
+			}
+		case "timeout":
+			var err error
+			result.Timeout, err = types.GetDurationValue(v)
+			if err != nil {
+				return result, fmt.Errorf("invalid timeout value: %w", err)
+			}
+		case "reflect":
+			var ok bool
+			result.UseReflectionProtocol, ok = v.(bool)
+			if !ok {
+				return result, fmt.Errorf("invalid reflect value: '%#v', it needs to be boolean", v)
+			}
+		case "maxReceiveSize":
+			var ok bool
+			result.MaxReceiveSize, ok = v.(int64)
+			if !ok {
+				return result, fmt.Errorf("invalid maxReceiveSize value: '%#v', it needs to be an integer", v)
+			}
+			if result.MaxReceiveSize < 0 {
+				return result, fmt.Errorf("invalid maxReceiveSize value: '%#v, it needs to be a positive integer", v)
+			}
+		case "maxSendSize":
+			var ok bool
+			result.MaxSendSize, ok = v.(int64)
+			if !ok {
+				return result, fmt.Errorf("invalid maxSendSize value: '%#v', it needs to be an integer", v)
+			}
+			if result.MaxSendSize < 0 {
+				return result, fmt.Errorf("invalid maxSendSize value: '%#v, it needs to be a positive integer", v)
+			}
+		case "tls":
+			if err := parseConnectTLSParam(result, v); err != nil {
+				return result, err
+			}
+		default:
+			return result, fmt.Errorf("unknown connect param: %q", k)
+		}
+	}
+
+	return result, nil
+}
+
+func parseConnectTLSParam(params *connectParams, v interface{}) error {
+	var ok bool
+	params.TLS, ok = v.(map[string]interface{})
+
+	if !ok {
+		return fmt.Errorf("invalid tls value: '%#v', expected (optional) keys: cert, key, password, and cacerts", v)
+	}
+	// optional map keys below
+	if cert, certok := params.TLS["cert"]; certok {
+		if _, ok = cert.(string); !ok {
+			return fmt.Errorf("invalid tls cert value: '%#v', it needs to be a PEM formatted string", v)
+		}
+	}
+	if key, keyok := params.TLS["key"]; keyok {
+		if _, ok = key.(string); !ok {
+			return fmt.Errorf("invalid tls key value: '%#v', it needs to be a PEM formatted string", v)
+		}
+	}
+	if pass, passok := params.TLS["password"]; passok {
+		if _, ok = pass.(string); !ok {
+			return fmt.Errorf("invalid tls password value: '%#v', it needs to be a string", v)
+		}
+	}
+	if cacerts, cacertsok := params.TLS["cacerts"]; cacertsok {
+		var cacertsArray []interface{}
+		if cacertsArray, ok = cacerts.([]interface{}); ok {
+			for _, cacertsArrayEntry := range cacertsArray {
+				if _, ok = cacertsArrayEntry.(string); !ok {
+					return fmt.Errorf("invalid tls cacerts value: '%#v',"+
+						" it needs to be a string or an array of PEM formatted strings", v)
+				}
+			}
+		} else if _, ok = cacerts.(string); !ok {
+			return fmt.Errorf("invalid tls cacerts value: '%#v',"+
+				" it needs to be a string or an array of PEM formatted strings", v)
+		}
+	}
+	return nil
 }
